refactor components formio
