+++ projects/valtimo/components/src/lib/components/form-io/components/form-io-currency/currency.component.html
