+++ projects/valtimo/components/src/lib/components/form-io/components/form-io-current-user/form-io-current-user.formio.ts
@@ -15,7 +15,7 @@
  */
 
 import {Injector} from '@angular/core';
-import {FormioCustomComponentInfo, registerCustomFormioComponent} from '../../../modules';
+import {FormioCustomComponentInfo, registerCustomFormioComponent} from '../../../../modules';
 import {formIoCurrentUserEditForm} from './form-io-current-user-edit-form';
 import {FormIoCurrentUserComponent} from './form-io-current-user.component';
