+++ projects/valtimo/components/src/lib/components/form-io/components/form-io-dummy/dummy.component.ts
@@ -15,25 +15,16 @@
  */
 
 import {Component, EventEmitter, Input, Output} from '@angular/core';
-import {FormioCustomComponent} from '../../../modules';
-import {KeycloakService} from 'keycloak-angular';
-import {KeycloakProfile} from 'keycloak-js';
+import {FormioCustomComponent} from '../../../../modules';
 
 @Component({
-  selector: 'valtimo-formio-current-user',
-  template: '{{ this.value?.firstName }} {{ this.value?.lastName }}<br />{{ this.value?.email }}',
-  styles: [],
+  selector: 'valtimo-dummy',
+  template: '',
   standalone: false,
 })
-export class FormIoCurrentUserComponent implements FormioCustomComponent<any> {
-  @Input() value: any;
-  @Input() disabled: boolean;
-  @Output() valueChange = new EventEmitter<any>();
-
-  constructor(private readonly keycloakService: KeycloakService) {
-    this.keycloakService.loadUserProfile().then((profile: KeycloakProfile) => {
-      this.value = profile;
-      this.valueChange.emit();
-    });
-  }
+export class FormioDummyComponent implements FormioCustomComponent<any> {
+  @Input() public value: string;
+  @Input() public disabled = false;
+  @Input() public required = false;
+  @Output() public valueChange = new EventEmitter<any>();
 }
