+++ projects/valtimo/components/src/lib/components/form-io/components/form-io-uploader/form-io-uploader.formio.ts
@@ -19,7 +19,7 @@ import {
   createCustomFormioComponent,
   FormioCustomComponentInfo,
   registerCustomFormioComponentWithClass,
-} from '../../../modules';
+} from '../../../../modules';
 import {FormIoUploaderComponent} from './form-io-uploader.component';
 import {formIoUploaderEditForm} from './form-io-uploader-edit-form';
