+++ projects/valtimo/components/src/lib/components/form-io/components/formio-value-resolver-selector/formio-value-resolver-selector.formio.ts
@@ -15,7 +15,7 @@
  */
 
 import {Injector} from '@angular/core';
-import {FormioCustomComponentInfo, registerCustomFormioComponent} from '../../../modules';
+import {FormioCustomComponentInfo, registerCustomFormioComponent} from '../../../../modules';
 import {formioValueResolverSelectorEditForm} from './formio-value-resolver-selector-edit-form';
 import {FormioValueResolverSelectorComponent} from './formio-value-resolver-selector.component';
 import {ValtimoWindow} from '@valtimo/shared';
