+++ projects/valtimo/components/src/lib/components/form-io/form-io.module.ts
@@ -16,11 +16,11 @@
 
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {FormioComponent} from './form-io.component';
-import {FormioBuilderComponent} from './form-io-builder.component';
+import {FormioComponent} from './components/form-io/form-io.component';
+import {FormioBuilderComponent} from './components/form-io-builder/form-io-builder.component';
 import {FormioAppConfig, FormioModule} from '@formio/angular';
 import {getFormioAppConfig} from './formio-config';
-import {FormIoUploaderComponent} from './form-io-uploader/form-io-uploader.component';
+import {FormIoUploaderComponent} from './components/form-io-uploader/form-io-uploader.component';
 import {DropzoneModule} from '../dropzone/dropzone.module';
 import {TranslateModule} from '@ngx-translate/core';
 import {DocumentModule} from '@valtimo/document';
@@ -28,14 +28,14 @@ import {FormIoDomService} from './services/form-io-dom.service';
 import {FileSizeModule} from '../file-size/file-size.module';
 import {ResourceModule} from '@valtimo/resource';
 import {RouterModule} from '@angular/router';
-import {FormIoCurrentUserComponent} from './form-io-current-user/form-io-current-user.component';
+import {FormIoCurrentUserComponent} from './components/form-io-current-user/form-io-current-user.component';
 import {ConfigService} from '@valtimo/shared';
-import {FormIoIbanComponent} from './form-io-iban/iban.component';
+import {FormIoIbanComponent} from './components/form-io-iban/iban.component';
 import {ReactiveFormsModule} from '@angular/forms';
-import {FormioValueResolverSelectorComponent} from './formio-value-resolver-selector/formio-value-resolver-selector.component';
-import {FormioDummyComponent} from './form-io-dummy/dummy.component';
+import {FormioValueResolverSelectorComponent} from './components/formio-value-resolver-selector/formio-value-resolver-selector.component';
+import {FormioDummyComponent} from './components/form-io-dummy/dummy.component';
 import {LayerModule} from 'carbon-components-angular';
-import {FormIoCurrencyComponent} from './form-io-currency/currency.component';
+import {FormIoCurrencyComponent} from './components/form-io-currency/currency.component';
 
 @NgModule({
   imports: [
