+++ projects/valtimo/components/src/public_api.ts
@@ -93,8 +93,8 @@ export * from './lib/components/data-list/data-list.module';
 
 export * from './lib/components/view-content/view-content.service';
 
-export * from './lib/components/form-io/form-io.component';
-export * from './lib/components/form-io/form-io-builder.component';
+export * from './lib/components/form-io/components/form-io/form-io.component';
+export * from './lib/components/form-io/components/form-io-builder/form-io-builder.component';
 export * from './lib/components/form-io/form-io.module';
 
 export * from './lib/components/modal/modal.module';
@@ -109,19 +109,19 @@ export * from './lib/components/searchable-dropdown/searchable-dropdown.componen
 export * from './lib/components/dropzone/dropzone.module';
 export * from './lib/components/dropzone/dropzone.component';
 
-export * from './lib/components/form-io/form-io-uploader/form-io-uploader.component';
-export * from './lib/components/form-io/form-io-uploader/form-io-uploader.formio';
+export * from './lib/components/form-io/components/form-io-uploader/form-io-uploader.component';
+export * from './lib/components/form-io/components/form-io-uploader/form-io-uploader.formio';
 
-export * from './lib/components/form-io/form-io-current-user/form-io-current-user.component';
-export * from './lib/components/form-io/form-io-current-user/form-io-current-user.formio';
+export * from './lib/components/form-io/components/form-io-current-user/form-io-current-user.component';
+export * from './lib/components/form-io/components/form-io-current-user/form-io-current-user.formio';
 
-export * from './lib/components/form-io/form-io-iban/iban.component';
-export * from './lib/components/form-io/form-io-iban/iban.formio';
+export * from './lib/components/form-io/components/form-io-iban/iban.component';
+export * from './lib/components/form-io/components/form-io-iban/iban.formio';
 
-export * from './lib/components/form-io/form-io-currency/currency.component';
-export * from './lib/components/form-io/form-io-currency/currency.formio';
+export * from './lib/components/form-io/components/form-io-currency/currency.component';
+export * from './lib/components/form-io/components/form-io-currency/currency.formio';
 
-export * from './lib/components/form-io/form-io-resource-selector/form-io-resource-selector.formio';
+export * from './lib/components/form-io/components/form-io-resource-selector/form-io-resource-selector.formio';
 
 export * from './lib/components/form-io/services/form-io-state.service';
 export * from './lib/components/form-io/services/form-io-dom.service';
@@ -159,11 +159,11 @@ export * from './lib/components/status-selector/status-selector.component';
 // value path selector
 export * from './lib/components/value-path-selector/value-path-selector.component';
 // formio value resolver selector
-export * from './lib/components/form-io/formio-value-resolver-selector/formio-value-resolver-selector.formio';
-export * from './lib/components/form-io/formio-value-resolver-selector/formio-value-resolver-selector.component';
+export * from './lib/components/form-io/components/formio-value-resolver-selector/formio-value-resolver-selector.formio';
+export * from './lib/components/form-io/components/formio-value-resolver-selector/formio-value-resolver-selector.component';
 // formio dummy
-export * from './lib/components/form-io/form-io-dummy/dummy.component';
-export * from './lib/components/form-io/form-io-dummy/dummy.formio';
+export * from './lib/components/form-io/components/form-io-dummy/dummy.component';
+export * from './lib/components/form-io/components/form-io-dummy/dummy.formio';
 // Table component
 export * from './lib/components/table/table.module';
 export * from './lib/components/table/table.component';
